retriver/requirements.txt
pymongo==4.14.1
kafka-python==2.2.15
pandas==2.3.2